chore: DRY scripts
