+++ scripts/lint-node.ts
@@ -1,7 +1,12 @@
 #!/usr/bin/env node
 
 import { readFile } from 'node:fs/promises';
-import { inspect, parseArgs, type ParseArgsOptionsConfig } from 'node:util';
+import { inspect } from 'node:util';
+
+import { ansi } from './utils/ansi.ts';
+import parseArgs, {
+  type ParseArgsOptionsWithDescription,
+} from './utils/parseArgs.ts';
 
 type NodeVersionSpec = string;
 
@@ -11,13 +16,6 @@ interface PackageJson {
   };
 }
 
-type ParseArgsOptionsWithDescription = Record<
-  string,
-  ParseArgsOptionsConfig[string] & {
-    description: string;
-  }
->;
-
 const options: ParseArgsOptionsWithDescription = {
   help: {
     type: 'boolean',
@@ -25,60 +23,6 @@ const options: ParseArgsOptionsWithDescription = {
     description: 'Show this help message',
   },
 };
-
-function isTruthyEnvVar(value: string | undefined): boolean {
-  return ['1', 'true', 'yes'].includes((value ?? '').toLowerCase());
-}
-
-function shouldUseColours(): boolean {
-  if (isTruthyEnvVar(process.env.FORCE_COLOR)) {
-    return true;
-  } else if (isTruthyEnvVar(process.env.NO_COLOR)) {
-    return false;
-  } else {
-    return process.stdout.isTTY && process.stderr.isTTY;
-  }
-}
-
-const ansi = shouldUseColours()
-  ? {
-      reset: '\u001b[0m',
-      bold: '\u001b[1m',
-      dim: '\u001b[2m',
-    }
-  : {
-      reset: '',
-      bold: '',
-      dim: '',
-    };
-
-function printHelp() {
-  interface Opt {
-    flags: string;
-    description: string;
-  }
-  const opts: Array<Opt> = Object.entries(options).map(([name, opt]) => {
-    const flags = [`--${name}`, opt.short ? `-${opt.short}` : null]
-      .filter(Boolean)
-      .join(', ');
-    return { flags, description: opt.description };
-  });
-  const maxFlagLength = Math.max(...opts.map((opt) => opt.flags.length));
-
-  process.stdout.write(
-    [
-      `${ansi.bold}Check if Node.js versions specs are in sync.${ansi.reset}`,
-      '',
-      `${ansi.bold}Usage:${ansi.reset} npm run lint:node ${ansi.dim}[OPTIONS]${ansi.reset}`,
-      '',
-      `${ansi.bold}Options:${ansi.reset}`,
-      ...opts.map(
-        (opt) => `  ${opt.flags.padEnd(maxFlagLength)}  ${opt.description}`,
-      ),
-      '',
-    ].join('\n'),
-  );
-}
 
 async function readFileRel(filePathRel: string): Promise<string> {
   const filePath = new URL(filePathRel, new URL('..', import.meta.url))
@@ -132,15 +76,12 @@ function lintNodeVersions(nvmrc: string, packageJson: PackageJson): boolean {
 }
 
 async function main() {
-  const { values } = parseArgs({
+  parseArgs({
+    description: 'Check if Node.js versions specs are in sync.',
+    invocation: 'npm run lint:node',
     options,
     allowPositionals: false,
   });
-
-  if (values.help) {
-    printHelp();
-    process.exit(0);
-  }
 
   let hasErrors = false;
